Refactored theme utils
#1504
