+++ packages/components/src/theme/ThemeCache.ts
@@ -6,7 +6,6 @@ import {
   DEFAULT_DARK_THEME_KEY,
   ThemeData,
   ThemePreloadData,
-  ThemePreloadStyleContent,
 } from './ThemeModel';
 
 const log = Log.module('ThemeCache');
@@ -16,74 +15,25 @@ export const THEME_CACHE_LOCAL_STORAGE_KEY = 'deephaven.themeCache';
 export type ThemeCacheEventType = 'change';
 
 /**
- * Creates a string containing preload style content for the current theme.
- * This resolves the current values of a few CSS variables that can be used
- * to style the page before the theme is loaded on next page load.
- */
-export function calculatePreloadStyleContent(): ThemePreloadStyleContent {
-  const pairs = ['--dh-accent-color', '--dh-background-color'].map(
-    key => `${key}:${getComputedStyle(document.body).getPropertyValue(key)}`
-  );
-
-  return `:root{${pairs.join(';')}}`;
-}
-
-/**
  * Cache containing registered themes.
  */
 export class ThemeCache {
-  constructor(private readonly localStorageKey: string) {
+  constructor(localStorageKey: string) {
+    this.localStorageKey = localStorageKey;
+
     bindAllMethods(this);
   }
 
-  readonly baseThemes: Map<string, ThemeData> = new Map();
+  private appliedThemes: [ThemeData] | [ThemeData, ThemeData] | null = null;
 
-  readonly customThemes: Map<string, ThemeData> = new Map();
+  private readonly baseThemes: Map<string, ThemeData> = new Map();
 
-  readonly eventListeners: Map<ThemeCacheEventType, Set<() => void>> =
+  private readonly customThemes: Map<string, ThemeData> = new Map();
+
+  private readonly eventListeners: Map<ThemeCacheEventType, Set<() => void>> =
     new Map();
 
-  appliedThemes: [ThemeData] | [ThemeData, ThemeData] | null = null;
-
-  /**
-   * Clear appliedThemes cache and emit a `change` event to any listeners.
-   */
-  onChange(): void {
-    log.debug(
-      'onChange:',
-      '\npreloadData:',
-      `themeKey: '${this.preloadData?.themeKey}', preloadStyleContent: '${this.preloadData?.preloadStyleContent}'`,
-      '\nbaseThemes:',
-      [...this.baseThemes.keys()].join(', '),
-      '\ncustomThemes:',
-      [...this.customThemes.keys()].join(', ')
-    );
-
-    this.appliedThemes = null;
-    this.eventListeners.get('change')?.forEach(handler => handler());
-  }
-
-  /**
-   * Register an event listener for the given event type.
-   * @param eventType The event type to listen for
-   * @param handler The handler to call when the event is emitted
-   * @returns A function to unregister the event listener
-   */
-  registerEventListener(
-    eventType: ThemeCacheEventType,
-    handler: () => void
-  ): () => void {
-    if (!this.eventListeners.has(eventType)) {
-      this.eventListeners.set(eventType, new Set());
-    }
-
-    this.eventListeners.get(eventType)?.add(handler);
-
-    /** Deregister the event handler. */
-    return () => {
-      this.eventListeners.get(eventType)?.delete(handler);
-    };
-  }
+  private readonly localStorageKey: string;
 
   private preloadData: ThemePreloadData | null = null;
 
@@ -91,7 +41,7 @@ export class ThemeCache {
    * Get the preload data from local storage or null if it does not exist or is
    * invalid
    */
-  getPreloadData(): Partial<ThemePreloadData> | null {
+  getPreloadData(): ThemePreloadData | null {
     if (this.preloadData == null) {
       const data = localStorage.getItem(this.localStorageKey);
 
@@ -112,31 +62,26 @@ export class ThemeCache {
   /**
    * Checks if given data matches the current preload data. If it does, do nothing.
    * If the data has changed, update the preload data and emit a `change` event.
-   * @param data
+   * @param preloadData The preload data to set
    */
-  setPreloadData(themeKey: string): void {
-    const newData = {
-      themeKey,
-      preloadStyleContent: calculatePreloadStyleContent(),
-    };
-
-    const newLocalStorageValue = JSON.stringify(newData);
+  setPreloadData(preloadData: ThemePreloadData): void {
+    const newLocalStorageValue = JSON.stringify(preloadData);
 
     if (newLocalStorageValue === localStorage.getItem(this.localStorageKey)) {
       return;
     }
 
-    this.preloadData = newData;
+    this.preloadData = preloadData;
     localStorage.setItem(this.localStorageKey, newLocalStorageValue);
 
     this.onChange();
   }
 
   /**
-   * Returns an array of the selected themes. The first item will always be one
+   * Returns an array of the applied themes. The first item will always be one
    * of the base themes. Optionally, the second item will be a custom theme.
    */
-  getSelectedThemes(): [ThemeData] | [ThemeData, ThemeData] {
+  getAppliedThemes(): [ThemeData] | [ThemeData, ThemeData] {
     if (this.appliedThemes == null) {
       const { themeKey } = this.getPreloadData() ?? {};
 
@@ -181,6 +126,24 @@ export class ThemeCache {
   }
 
   /**
+   * Clear appliedThemes cache and emit a `change` event to any listeners.
+   */
+  onChange(): void {
+    log.debug(
+      'onChange:',
+      '\npreloadData:',
+      `themeKey: '${this.preloadData?.themeKey}', preloadStyleContent: '${this.preloadData?.preloadStyleContent}'`,
+      '\nbaseThemes:',
+      [...this.baseThemes.keys()].join(', '),
+      '\ncustomThemes:',
+      [...this.customThemes.keys()].join(', ')
+    );
+
+    this.appliedThemes = null;
+    this.eventListeners.get('change')?.forEach(handler => handler());
+  }
+
+  /**
    * Register the given base themes with the cache.
    * @param themeDatas The base themes to register
    */
@@ -209,11 +172,25 @@ export class ThemeCache {
   }
 
   /**
-   * Select the theme with the given key.
-   * @param themeKey The theme key to select
+   * Register an event listener for the given event type.
+   * @param eventType The event type to listen for
+   * @param handler The handler to call when the event is emitted
+   * @returns A function to unregister the event listener
    */
-  setSelectedTheme(themeKey: string): void {
-    this.setPreloadData(themeKey);
+  registerEventListener(
+    eventType: ThemeCacheEventType,
+    handler: () => void
+  ): () => void {
+    if (!this.eventListeners.has(eventType)) {
+      this.eventListeners.set(eventType, new Set());
+    }
+
+    this.eventListeners.get(eventType)?.add(handler);
+
+    /** Deregister the event handler. */
+    return () => {
+      this.eventListeners.get(eventType)?.delete(handler);
+    };
   }
 }
