+++ packages/components/src/theme/index.ts
@@ -2,3 +2,5 @@ export * from './ThemeCache';
 export * from './ThemeContext';
 export * from './ThemeModel';
 export * from './ThemeUtils';
+export * from './useInitializeThemeContextValue';
+export * from './useTheme';
