+++ packages/components/src/theme/useTheme.ts
@@ -1,0 +1,14 @@
+import { useContextOrThrow } from '@deephaven/react-hooks';
+import { ThemeContext, ThemeContextValue } from './ThemeContext';
+
+/**
+ * Hook to get the current `ThemeContextValue`.
+ */
+export function useTheme(): ThemeContextValue {
+  return useContextOrThrow(
+    ThemeContext,
+    'No ThemeContext value found. Component must be wrapped in a ThemeContext.Provider'
+  );
+}
+
+export default useTheme;
